On some systems u_char* (pointer) is a unsigned long data type and not an unsigned integer.  Because getEndOfData returns u_char* and because data is u_char*, then, on those systems there is some ambiguity on which fxmin function is to be used (the one for u_int or the one for u_long).
Since stripSize is u_int, it therefore doesn't make any sense to be
subtracting a long data type from an int data type.

So this typing change (specifying that getEndOfData - data is to yield a
u_int result) fixes both of these ambiguites and allows faxq to build
properly on systems with long data types for u_char*.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@2068 5505949e-d877-4686-9e03-c53b7a51b376

+++ faxd/faxQueueApp.c++
@@ -1,4 +1,4 @@
-/*	$Id: faxQueueApp.c++ 949 2009-10-17 21:20:59Z faxguy $ */
+/*	$Id: faxQueueApp.c++ 950 2009-10-27 23:29:30Z faxguy $ */
 /*
  * Copyright (c) 1990-1996 Sam Leffler
  * Copyright (c) 1991-1996 Silicon Graphics, Inc.
@@ -1029,7 +1029,7 @@ faxQueueApp::preparePageChop(const FaxRequest& req,
 	if (dec.getLastBlanks() > minRows)
 	{
 	    pagehandling.append(fxStr::format("Z%04x",
-		fxmin((unsigned)0xFFFF, stripSize - (dec.getEndOfPage() - data))));
+		fxmin((unsigned)0xFFFF, stripSize - (u_int)(dec.getEndOfPage() - data))));
 	}
     }
     delete [] data;
